requirements.txt
nicegui
pydantic 
fastapi 
uvicorn
pdfplumber 
pytesseract 
pillow
spacy
rapidfuzz
fitz 
